+++ test/cases/msg-exception-loader-preprocessor-unsupported/webpack.config.js
@@ -1,0 +1,18 @@
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+  devtool: false,
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: './src/index.html',
+      },
+      loaderOptions: {
+        // test exception unsupported preprocessor
+        preprocessor: '_unsupported_',
+      },
+    }),
+  ],
+};